[LOCAL] Relax unaligned access assertion when type is byte aligned
This commit has been cherry-picked from upstream LLVM review D39946.
Once that patch lands in LLVM trunk, we should revert this commit and
cherry-pick the official one.

Original message:
-----------------

This relaxes an assertion inside SelectionDAGBuilder which is overly
restrictive on targets which have no concept of alignment (such as AVR).

In these architectures, all types are aligned to 8-bits.

After this, LLVM will only assert that accesses are aligned on targets
which actually require alignment.