AMDGPU: Process SDWA block at a time
Right now this loops over the entire function every time there
is a change, which is not very efficient. There's no practical
reason to track this so globally, since the code motion optimization
passes should be sinking instructions with single uses and
the pass currently will not fold with multiple uses.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@324667 91177308-0d34-0410-b5e6-96231b3b80d8