[X86] In lowerVSELECTtoVectorShuffle, don't map undef select condition to undef in shuffle mask.
Undef in select condition means we should pick the element from one side or the other. An undef in a shuffle mask means pick any element from either source or worse.

I suspect by the time we get here most of the undefs in a constant vector have been removed by other things, but doing this for safety.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@325394 91177308-0d34-0410-b5e6-96231b3b80d8