[PDB] Print some more details when explaining MSF fields.
When we determine that a field belongs to an MSF super block or
the free page map, we wouldn't print any additional information.

With this patch, we now print the value of the field (for super
block fields) or the allocation status of the specified byte (in
the case of offsets in the FPM).

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@328808 91177308-0d34-0410-b5e6-96231b3b80d8