[llvm-pdbutil] Add the ability to explain binary files.
Using this, you can use llvm-pdbutil to export the contents of a
stream to a binary file, then run explain on the binary file so
that it treats the offset as an offset into the stream instead
of an offset into a file.  This makes it easy to compare the
contents of the same stream from two different files.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@329207 91177308-0d34-0410-b5e6-96231b3b80d8